- ApplicationManager class created for TestBase to delegate all methods to it
