+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/GroupCreationTests.java
@@ -7,11 +7,11 @@ public class GroupCreationTests extends TestBase {
   @Test
   public void testGroupCreation() {
 
-    goToGroupPage();
-    initGroupCreation();
-    fillGroupForm(new GroupData("test1", "test2", "test3"));
-    submitGroupCreation();
-    returnToGroupPage();
+    app.goToGroupPage();
+    app.initGroupCreation();
+    app.fillGroupForm(new GroupData("test1", "test2", "test3"));
+    app.submitGroupCreation();
+    app.returnToGroupPage();
 
   }
