+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/GroupDeletionTests.java
@@ -7,10 +7,10 @@ public class GroupDeletionTests extends TestBase {
   @Test
   public void testGroupDeletion() {
 
-    goToGroupPage();
-    selectGroup();
-    deleteSelectedGroups();
-    returnToGroupPage();
+    app.goToGroupPage();
+    app.selectGroup();
+    app.deleteSelectedGroups();
+    app.returnToGroupPage();
 
   }
